Выносит createDestinationPhotosTemplate из утилит в темлпейт
