+++ src/view/event-edit-view.js
@@ -2,6 +2,7 @@ import he from 'he';
 import AbstractStatefulView from '../framework/view/abstract-stateful-view.js';
 import createTypeListTemplate from '../templates/type-list-template.js';
 import createDestinationListTemplate from '../templates/destination-list-template.js';
+import createDestinationPhotosTemplate from '../templates/destination-photos-template.js';
 
 import {
   humanizePointInputDateTimeType,
@@ -9,7 +10,6 @@ import {
   getTypeOffers,
   getDestinationPhotos,
   getDestinationObject,
-  createDestinationPhotosTemplate,
   createDestinationDescriptionTemplate,
   createOffersSectionTemplateTemplate,
 } from '../utils/point.js';
